Patch unable to initiate new vote
A small fix that should patch being unable to start a new vote.

Added a secondary check to see if the created vote should be 'expired' by now, and if it is then allow for a new vote to be created.

Since this bug prevents players from creating new votes my intention is to forcefully check and see if the cooldown should be done.